Add --checksum/--no-checksum flag
Option exists because the dbip checksum is retrieved via screen
scrape. If the page changes, checksum validation will fail. In that
case, the --no-checksum option can be used to workaround the issue
until a fix is available.

The flag applies to the MaxMind provider as well for consistency.

Minor refactoring of maxmind.py:
 - use requests instead of urllib for downloads
 - pass file to be checksum'd all at once instead of 4K chunks